Disable webview bouncing
When the webview's scrollview is allowed to bounce it sometimes scrolls
to its bouncing margins when the user intended to scroll the page loaded
in the web view. This results in ugly and annoying effects such as the
top bar of the web interface scrolling down from the status bar. Since
the loaded page has a bouncing effect anyway it seems natural to disable
this capability.