Fix always registering every sensor on not_registered
When trying to register only a given list of sensors, all of the Promise are created immediately, and the webhook response promises invoke immediately. In the end, all the filtering was doing was preventing waiting for all of them, thus making the DoS in an infinite loop even stronger!

This modifies the filter check to do the filter before creating the promises, and thus only will register a sensor that is failing.