Validate manually-entered URL strings
If you, for example, enter `127.0.0.1:8123` without the scheme, this ends up being a quasi-valid URL that will fail later down the pipeline.

If you enter something that's even closer to valid like `http:127.0.0.1:8123` it will crash trying to make a WebSocket connection, since that API strictly looks for http/https schemes.

This prompts the user if they forget a scheme, trims any whitespace, and makes sure that the thing passes an is-URL heuristic (by initting a URL with the string) before allowing the user to continue.