Fix crash on discovery from 0.110 (#615)
Amusingly enough, this is actually a bug in Foundation. When we retrieve the  `txtRecordData` dictionary representation, we're supposed to get back a `[String: Data]`. However, Foundation is more than happy to return an `NSNull` instead of the `NSData` which causes the bridging of the dictionary into Swift to crash.

Works around this problem by not allowing the bridge into Swift to type the value and instead filters out the invalid ones. I can confirm that I can still connect to a 0.110 instance after this filtering.

Fixes (iOS-side) #590.