Add Location & Motion permission to settings
Without an entry point in the app to force the location/motion manager to request permission, a user who didn't enable it during onboarding won't have a way to turn it on without logging out or deleting the app.

This should also reduce the confusion about the current status by allowing the user to jump into Settings to e.g. change from Never to Always.

This also needs to be done for Notifications, but that'll be a different PR.