State Restoration: Save Last Page
Saves state (right now just the page the WebView is displaying) when backgrounding the app normally. This handles a few weird edge cases like saving a dashboard which ends up being deleted since the last restoration.

This also fixes a bug where the WebView was being told to load the frontend twice on startup, which explains some of the logo flashing that occurs on a clean start.

Adds an option (defaulting to on) to settings which controls whether state is restored.