Sensor List: Show Last Sent Versions
Caches the last sensor update we sent, which gives a better indication of what's happening. Includes the date in the footer that it was done. Bonus: this now also includes the Geocoded sensor in the sensor list.

This makes all connects also trigger location updates, too, so Periodic can actually update that value if it is e.g. a wall tablet that's never going to move or update otherwise.