ZoneManager: Directly observe Realm for syncing
Instead of needing to wire up locations that are mutating zones into the manager, this makes the manager observe for any changes and sync them as a result.

One tough thing here: Realm, for some unbelievable reason, notifies about changes in the transaction that is mutating them. This means that anything which is observing the changes and may perform its own write in response needs to make sure it doesn't start a new write transaction, which also unbelievably crashes.