Add quality thresholds to location results
This requires locations we're considering as updates to:

- Not have a longitude or latitude of exactly 0. I'm not sure why this is happening, and it almost certainly is not something we're doing (I can't find any reason why it would be) but at least we can detect it. Fixes #684.
- Not have an accuracy that exceeds 1500m or be older than 600s. Refs #298. These are numbers that can be tuned in the future, too.

The tradeoff here is that we may be disregarding poor location information when entering and exiting monitored regions. We may need to write some compensating "zone enters/exits are not location-based" logic if this proves to be the case.

This also bumps up our requested accuracy level from 200m to "Best." I'm guessing in practice there may be no difference in the poor results iOS seems to be sending us.