Extend local logging from ~1.5 hours to ~3 days
The default values are 10 logs at 600 seconds each with a max file size of 1.5mb, so depending on how active the app is, anything that happened earlier in the day is gone. This extends our maximum logging size from 15mb to 30mb and a much longer duration of time.

Limits are about doubled for TestFlight builds.