Webhooks: expire requests after 2 hours, add logging
By default, background networking requests will stick around and continually retry for 7 days. However, if the user's configuration for some reason has requests point to a URL that will always time out, this means we'll have a bunch of requests sitting around doing nothing. This sets the maximum time we're wiling to have background retries to 2 hours.

Adds a log on startup to see how many tasks are currently in the queue for the session, in case we have a backup. Mostly helpful for forensic investigation.