Handle deleting and reinstalling the app better
When the app is deleted, but the keychain values (webhook, auth token) persist, our state looks like:

1. The API, WebhookManager, etc., are all alive and well, talking to the server.
1. The UI shows the onboarding flow.

This is because we control whether onboarding should be displayed based on a UserDefaults entry for whether they've finished it, which is less likely than the keychain to persist across installs. What happens next is:

1. A webhook update is sent, since we're launching
1. If the webhook already existed, it's updated. If it did not, using the auth token that is still around, we re-register and create a new integration.
1. The user finishes onboarding, which creates a _new_ auth token, and registers a new integration.

If the integration that was still around in the keychain was for the same `identifierForVendor` (what we send for `device_id`) this doesn't do anything surprising - it just updates the sensors and registration.

If the integration that was still around, however, is for a different (or none) `device_id` this will create a new integration. Giving us a total of 2 integrations for 1 install.

Fixes #819 by using the keychain values we discover on startup to decide if we're done with onboarding.

I tested 2 scenarios:

1. The keychain values are valid, in which case the app comes back alive and works normally, albeit with most settings reset.

This is fine now because all of the onboarding-related settings are now available in-app; before that was the case, skipping onboarding was potentially bad for location, etc., permission.

2. The keychain values are invalid. When this happens we show the normal app UI until the server rejects our credentials, so it behaves much like auth tokens being revoked on startup.