Add Perform Action intent, donate intents & suggestions
Adds a new Shortcut action called "Perform Event" which performs the given event. The parameter to this Shortcut requires the user choose from a list which action to execute.

This is donated as an intent every time an action is performed, and this intent is flagged as suggest-able, so it'll start appearing on lock screen, search, and in the Shortcuts app.