Allow "redundant" zone events to trigger location updates
This is only an issue for zones <100m, and is a logic issue following adding additional monitored regions for smaller zones. Since we are now giving 3 regions to smaller zones, we can't expect a direct "inside region = inside zone" relationship; you can be in 1 of the regions and outside the others but still be in the zone.

What this means is that a user can enter one of the wrapping regions, have their location (albeit outside the zone) updated from there and then when they enter one of the overlapping regions the app chooses not to update the location since it believes they are already inside the zone.

This fixes the issue by not caring if the user is in a zone or not when deciding whether to update location.