Improve Connection Settings saving and editing
Moves setting an internal or external URL into a sub-screen which has cancel/save buttons and an explicit lifecycle around whether the save is successful:

- When successful, pops off and shows the new value.
- When unsuccessful for connectivity reasons, allows the user to ignore the error and persist the URL change anyway.
- When unsuccessful because we reject the URL as being Nabu Casa or trying to clear the last one, we tell the user we can't save.

This also allows clearing external if internal is set and fixes trying to reset to external when there is no external.

Improves how and when we persist SSIDs in the list to be the same save button next to internal URL and adds a warning if location permission suggests we won't be able to determine SSID information.

Fixes #717.