Fix infinite loop syncing Actions to watch
This happens because the Mappable-variant encode/decode of an Action requires a write transaction (since ObjectMapper's mapping is `inout`, which triggers the "only set in a write transaction" check in Realm) which then, directly, caused a new update to the Watch.

Worked around this by making Action and WatchComplication sync via ImmutableMappable which does not do anything inout, and thus doesn't require a write transaction.

Fixes a few more issues:
- Setting shortcutItems and syncing to watch is now done entirely in the Actions observer; removed from Settings.
- Fix iPhone-not-connected action sending for Scene-synced Actions, which don't have their RLMScene synced to the watch.
- Fixed the empty state label not hiding if actions go in or out of 0-count after initially.