Fix crash customizing a just-created Scene Action
When we were persisting the Action created by the Scene, if the Scene wasn't already added to Realm then we'd drop the Action instead of persisting it. This is because we're not directly referencing the Action in the persisted Scene -- we rely on the automatic inverse relationship list for accessing it. Resolves the crash by making sure we always use an active Realm (rather than an optional one) when persisting the Action.

Fixes #885.