Fix crash on iOS 12 (by weak-linking CoreNFC)
iOS 13 ships with CoreNFC.framework, but iOS 12 (and presumably earlier, where it _is_ declared as available) has the framework on some devices but not others. Big thanks to @RobC in Discord for pulling an old iPad Air out of his closet to install a test build for verifying this fixes it.