Fix database error for duplicate notification categories (#897)
I have no idea how it's possible to get the database in this state. I believe it's going to boil down to an older version of Realm allowing it to happen, and a newer version of Realm finding it to be a problem.

The UI currently allows you to create a second instance of the same identifier (but Realm will collapse them into the same object, confusingly enough) so that's possibly one of the causes, and the other could be old import code for the categories.

Either way: this will delete one of the categories on disk that has the same identifier as another.