Scene/Multi-Window support for iOS 13+ / Mac (#950)
Migrates from AppDelegate-based to SceneDelegate-based APIs on iOS/iPadOS/macCatalyst 13+. This involves a few major refactors:

- URL handling is now done in the `IncomingURLHandler` which is explicitly owned by the SceneDelegate. A compatibility wrapper for iOS 12 and below is used to keep the AppDelegate hooks working.
- WebViews gain a WindowController, which are responsible for deciding view controllers that appear in a particular window. This allows us to support putting together WebView-based windows in various places (iOS 12 without scenes, iOS 13+ with scenes).

This gives us multi-window support on iPadOS 13+ and Mac Catalyst 13+. For iPadOS, you can drag the app icon to create a secondary+ window as if you're launching a new app. On Mac, File > New will create a new window. On both, opening Settings will happen in a new window, and on Mac it will also open a new window for About.

The iPhone experience remains unchanged as it does not support multiple windows.