Add 'Microphone' sensors, like the 'Camera' ones (#986)
Refactors the Camera sensor code into general Input sensors, which also watch for Microphone in use status.

This uses the CoreAudio APIs which are _incredibly_ similar to the CoreMediaIO APIs for accessing devices. They're similar enough that the code paths can be shared, but dissimilar enough that e.g. getting property data has different method signatures so we do need to pay attention to which one we're using.

Like Cameras, these live-update when their in-use status changes and they automatically update as devices are attached or detached from the system.

If e.g. a camera and microphone have the same name, this suffixes them like "Razer Kiyo (Camera)" and "Razer Kiyo (Microphone)".