Complication Configuration: Explicit cancel/save, fix initial values
Although the complication logic was persisting the information correctly, the initial load of the view controller wasn't showing the right values -- likely due to some lifecycle in the table views. Fixes #752 by moving the hidden/not-hidden state of various sections into the sections themselves and asks them to update. This feels a bit less fragile and certainly more explicit.

This moves to having an explicit cancel button (which doesn't persist) and save button (which does) like the action/notification category editing. This also doesn't mutate the config while changing templates, so cancel is always an escape button.