Downgrade Realm to 5.3.5 & fix fatal alert, update Cocoapods (#1022)
Realm 5.3.6 has a workaround for the iOS 14 b3-b6 (but reverted by Apple in b7) crash. However, this workaround also causes deadlocks/crashes restoring form the background on iOS 12 and below. This reverts to the build before the regression, which will cause us to crash on the previously-noted iOS 14 betas, but will make iOS 12 and 11 work as expected.

This also fixes the alert we present when the Realm can't be be set up, which broke for Scenes.

Fixes #1020.