Fix not-fully-defined Shortcut execution by asking for values
When we use `.confirmationRequired(…)` for the result of a parameter resolve, the system drops it and silently stops executing the shortcut (without erroring or success). This is a change in behavior from iOS 13.

Working around this improves the user experience anyway - if a value is required, for example, telling the system this allows it to prompt the user for a value rather than erroring.

Fixes #1031.