Share Extension which fires `mobile_app.share` (#1014)
When sharing text or a URL, the Home Assistant share extension shows up as an option. When performing the action, this sends an event named `mobile_app.share` which looks like:

```json
{
   "entered" : "text entered into share extension",
   "text" : "text if provided",
   "url" : "https://url/if/provided"
}
```

This does not support sharing files or images. We could extend it to do so, but we need to figure out a clean way to handle it; sending e.g. base64-encoded file data seems messy.

Fixes #788.