Add Mac SSID/BSSID sensors
Uses the new MacBridge to talk to macOS-only CoreWLAN, which can give us the WiFi information about the Mac.

- Adds back in the SSID/BSSID sensors, mirroring how iOS creates them. If there are no WiFi devices, this doesn't add any sensors.
- Restores the Internal URL when WiFi is available.

This does not include the connectivity type sensors, because they do not provide useful information as of yet -- if we can say e.g. which ethernet devices are connected, that might be more useful.