Add Mac SSID/BSSID sensors (#1070)
Uses the new MacBridge to talk to macOS-only CoreWLAN, which can give us the WiFi information about the Mac.

- Adds back in the SSID/BSSID sensors, mirroring how iOS creates them. If there are no WiFi devices, this doesn't add any sensors.
- Restores the Internal URL when WiFi is available.

This is intentionally (currently) ignoring the fact that there may be more than one WiFi interface, because I did not want to refactor the connection things throughout the app yet.

This does not include the connectivity type sensors, because they do not provide useful information as of yet -- if we can say e.g. which ethernet devices are connected, that might be more useful.

This framework _does_ support live-updating us if there are any changes, so an update signaler for this may be good. Or it may conflict with other things we do on reachability changes, we'll see.

Fixes #941.