Update Action to download localized strings (#1094)
Downloads and commits localized strings, only if the English strings don't change as a result. This avoids the situation where we do not have the latest copy of English uploaded.

If we want to update the English strings via Lokalise, this will require a manual commit, or maybe parameterizing this Action or something in the future.