Add "Update Sensors" intent
This is functionally nearly identical to the "Send Location" intent, but doesn't require providing a location to the shortcut to execute, nor does it need to do a network request to geocode, so it should be a lot faster at providing updates.

This is useful for e.g. "When Low Power Mode is turned on or off" or "When (phone) is connected or disconnected from power" automations in Shortcuts.app.

I submitted Intents.strings to Lokalise externally so it should come down in the next download into English as well.