Don't trigger location update on initial zone monitoring
This avoids sending `ios.zone_{entered,exited}` on startup, and also stops resetting which monitored regions we are monitoring as a result of a location update, which may have been caused by resetting which monitored regions we are monitoring, etc., etc.

Since we do not locally need the in/out state of Zones, we can deal with their status being eventually correct instead of initially or whenever they change.