Fix non-English Gauge/Ring, fix Column 2 Alignment
These styles were using the localized strings as options, which were then not equating. The way the current getters work, their failure to exist in the dictionary it would always use 'ring' for Gauge and 'closed' for Ring.

Replaces Row 2 Alignment (which had a similar issue) with Column 2 Alignment, which is what the complications want.