Handle ending the WatchConnectivity background task when no data is left
Apple's docs say:

> Background data may be received before, after, or during your handle(_:) method. The handle(_:) may also receive multiple background watch connectivity tasks at once, may be called while you’re currently processing another set of background watch connectivity tasks, or may even be called while your app is in the foreground. You must check for and handle all of these cases.

Although we check if there's more data before terminating existing tasks, we don't check if we've already handled the data being requested. This means we could get in a situation where we process the data before we're told to handle the background task.

This is, in other words, the WatchConnectivity version of #1203.