Add "Placeholder" complication
Since we only offer complications for configured versions, we don't have a quick "launch the app" complication available without creating one, and you cannot create one that's just the HA logo.

Rather than adding a new template configuration for the app icon, this adds a complication to all places we support which show some variant of "Home Assistant", "HA" and the app icon without any kind of "this is an error!" state.